WIP: add ability to use Keep-A-Changelog to set release date in CHANGELOG and fetch CHANGELOG contents
This patch modifies the workings of the `ReleaseCommand` slightly.

Instead of depending on the `CreateReleaseText` interface, it now
depends on a new `Laminas\AutomaticReleases\Changelog\ReleaseChangelogAndFetchContents`
interface. This new interface consumes a `ReleaseChangelogEvent`, which
stores a variety of runtime-determined values (console input/output,
milestone, repository name and path, etc.). Implementors of this
interface pull values from the event that they might use to fetch or
create the changelog and return it. If they are incapable of doing so,
they return null.

I have created three implementations:

- `UseKeepAChangelogEventsToReleaseAndFetchChangelog`, which checks to
  see if there is a `CHANGELOG.md` file in the repository. If not, it
  returns null immediately. Otherwise, it sets the release date in the
  changelog for the given version, commits the file (more on that
  later), and pushes the changes to the release branch. It then returns
  the discovered changelog contents

- `CreateChangelogViaMilestone` pulls data from the event for use with
  the original `CreateReleaseText` implementation, and decorates such an
  instance.

- `ReleaseChangelogAndFetchContentsAggregate` takes a list of
  `ReleaseChangelogAndFetchContents` instances, and returns the first
  changelog discovered. If none return one, it generates a basic one using
  just the repository name and release version.

I have updated `bin/console.php` to create a
`ReleaseChangelogAndFetchContentsAggregate` instance composing each of
the other two types, with the keep-a-changelog version being first.

To assist in how the `UseKeepAChangelogEventsToReleaseAndFetchChangelog`
implementation works, I created `Laminas\AutomaticReleases\Git\CommitFile`,
which describes a commit operation on a single file. `Laminas\AutomaticReleases\Git\CommitFileViaConsole`
does a `git add` operation followed by a `git commit` operation on the
file.

Currently, no tests; I needed to play with architecture to figure out a
workable approach first.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>