refactor: idempotent creation of release text via keep-a-changelog
Updates `CreateReleaseText` to:

- Add a `BranchName $sourceBranch` argument
- Add a `string $repositoryDir` argument
- Add a new `canCreateReleaseText()` method with identical arguments,
  returning a boolean. The method should be called before invoking the
  implementation to create release text.

After making those changes, updated `CreateReleaseTextThroughChangelog`
to add the arguments and new method, and added a test for the new
method.

Finally, adds a new class, `CreateReleaseTextThroughChangelog`, which
uses the phly/keep-a-changelog `ChangelogParser` to determine if (a) it
can locate a CHANGELOG.md file and changelog entry matching the release
version, and then (b) pull it and return it.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>