feat: adds psr/log and monolog/monolog as dependencies
Adds psr/log and monolog/monolog as dependencies. Additionally, sets up
the environment and application to allow consuming a log.

The new `LOG_LEVEL` environment variable is optional, and will default
to `'INFO'`. In the `bin/console.php` file, we now create a Monolog
logger instance with a `StreamHandler` that writes to the `STDERR`
stream, using the configured `LOG_LEVEL`.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>