feat: functionality for bumping, commiting, and pushing new changelog version entries
This patch introduces the interface `BumpAndCommitChangelogVersion`,
describing an operation where a new changelog entry is injected into the
changelog file, if it exists; the new version is determined by the first
argument passed, which must be one of a set of known constant values
describing patch and minor release bumps.

This patch also introduces the `CheckoutBranch` interface, describing
how to checkout a specific branch of a repository. It has one
implementation, `CheckoutBranchViaConsole`.

The new class `BumpAndCommitChangelogVersionViaKeepAChangelog` composes
a `CheckoutBranch` instance, as well as `CommitFile`, `Push`, and
`LoggerInterface` instances, to assist in the work of updating the
changelog. If no changelog is detected in the branch, no work is done.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>