fix: do not checkout the branch passed to `CommitFileViaConsole`
We need to assume we're on the correct branch; otherwise, a checkout
operation will lead to an error, aborting the operation.

Instead, I've added functionality to verify that we are on the branch
provided; if not, an exception is raised.

The changes also required annotating that the repository directory,
filename, and commit message arguments could not be empty, which
required changes elsewhere to ensure this is true before calling on
`CommitFile`.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>