fix: check for changelog and/or pull contents from changelog using remote origin
As it turns out, while you can use `git show {ref}:{filename}` to both
test for the existing of a file at a given reference, as well as display
its contents, it only works if the reference is present.

When the github action starts, it _exports_ the default branch to the
working directory. Our code then _fetches_ from the remote, which then
sets the current branch to the _default_ branch, which may or may not be
the branch we are releasing from. It's not until we do a checkout
operation that the release branch can be guaranteed as a reference in
the checkout.

The reference _does_ exist in the `origin` remote, and since we have all
history, that means the content is present in our checkout; it's just
that the reference doesn't exist **IN** the checkout.

As such, for idempotent operations where we are checking for the
existence of a file, or pulling its contents, we need to qualify the
reference with the remote name (e.g., `{remote}/1.0.x`). Since we can
guarantee that the `origin` remote is present, we can use that remote
name.

Fiexes #51

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>