Use `git config credential.helper 'store --file=/tmp/file' for `git push`
This should fix some underlying design issues described in #198

The full URI of the repository, which includes a token in plaintext, is now no
longer the actual `git remote` URI, and is instead stored in a temporary
file which `git` uses to authenticate with the remote.

This prevents the remote URI from leaking in `STDOUT`/`STDERR`, which are
visible in case of crashes.